[#66] Move button helper to App link directory
